+++ .github/workflows/django-test.yml
@@ -2,7 +2,7 @@ name: Django Test
 
 on:
   push:
-    branches: [ master, dev, lyh543 ]
+    branches: [ master, dev ]
   pull_request:
     branches: [ master, dev ]
