+++ coverage-test.bat
@@ -1,0 +1,3 @@
+coverage run --source=./ manage.py test --noinput
+coverage html
+start htmlcov/index.html
