+++ uestcmsc_webapp_backend/urls.py
@@ -29,5 +29,6 @@ urlpatterns = [
     url(r'^docs/$', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
     path('admin/', admin.site.urls),
     path('users/', include('users.urls')),
+    path('accounts/', include('accounts.urls')),
     path('', index)
 ]
